# go/model_auth_result.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type AuthResult string

// List of AuthResult
const (
	SUCCESS AuthResult = "AUTHENTICATION_SUCCESS"
	FAILURE AuthResult = "AUTHENTICATION_FAILURE"
	ONGOING AuthResult = "AUTHENTICATION_ONGOING"
)