# go/model_authentication_info.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type AuthenticationInfo struct {

	SupiOrSuci string `json:"supiOrSuci"`

	ServingNetworkName string `json:"servingNetworkName"`

	ResynchronizationInfo *ResynchronizationInfo `json:"resynchronizationInfo,omitempty"`

	TraceData *TraceData `json:"traceData,omitempty"`
}