# go/model_av5g_aka.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Av5gAka struct {

	Rand string `json:"rand"`

	HxresStar string `json:"hxresStar"`

	Autn string `json:"autn"`
}