# go/model_confirmation_data.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type ConfirmationData struct {

	ResStar string `json:"resStar"`
}