# go/model_confirmation_data_response.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type ConfirmationDataResponse struct {

	AuthResult *AuthResult `json:"authResult"`

	Supi string `json:"supi,omitempty"`

	Kseaf string `json:"kseaf,omitempty"`
}