# go/model_eap_session.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type EapSession struct {

	// contains an EAP packet
	EapPayload string `json:"eapPayload"`

	KSeaf string `json:"kSeaf,omitempty"`

	Links map[string]LinksValueSchema `json:"_links,omitempty"`

	AuthResult *AuthResult `json:"authResult,omitempty"`

	Supi string `json:"supi,omitempty"`
}