# go/model_invalid_param.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type InvalidParam struct {

	Param string `json:"param"`

	Reason string `json:"reason,omitempty"`
}