# go/model_links_value_schema.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type LinksValueSchema struct {

	Href string `json:"href,omitempty"`
}