# go/model_resynchronization_info.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type ResynchronizationInfo struct {

	Rand string `json:"rand"`

	Auts string `json:"auts"`
}