# go/model_trace_data.go
/*
 * AUSF API
 *
 * OpenAPI specification for AUSF
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type TraceData struct {

	TraceRef string `json:"traceRef"`

	TraceDepth *TraceDepth `json:"traceDepth"`

	NeTypeList string `json:"neTypeList"`

	EventList string `json:"eventList"`

	CollectionEntityIpv4Addr string `json:"collectionEntityIpv4Addr,omitempty"`

	CollectionEntityIpv6Addr string `json:"collectionEntityIpv6Addr,omitempty"`

	InterfaceList string `json:"interfaceList,omitempty"`
}